Apply improvements to metric
